change long long to uint64_t

+++ sudoku.h
@@ -43,7 +43,7 @@ void setDigitAt(Sudoku* sudoku, const uint32_t sudokuNo, uint32_t row, uint32_t 
 
     if (row < 6) {
         sudoku->rows[sudokuNo][row] &= ~(0xFULL << (20 + 4 * col));
-        sudoku->rows[sudokuNo][row] |= (long long)digit << (20 + 4 * col);
+        sudoku->rows[sudokuNo][row] |= (uint64_t)digit << (20 + 4 * col);
         return;
     }
 
@@ -56,7 +56,7 @@ void setDigitAt(Sudoku* sudoku, const uint32_t sudokuNo, uint32_t row, uint32_t 
     }
 
     sudoku->rows[sudokuNo][row] &= ~(0xFULL << (4 * col));
-    sudoku->rows[sudokuNo][row] |= (long long)digit << (4 * col);
+    sudoku->rows[sudokuNo][row] |= (uint64_t)digit << (4 * col);
 }
 
 void printSudoku(const Sudoku* sudoku, const uint32_t sudokuNo) {
